src/requirements.txt
flask>=3.1.1
flask-Markdown>=0.3
Bootstrap-Flask>=2.5.0
omegaconf>=2.3.0
rich>=14.1.0
mistune>=3.1.3
python-frontmatter>=1.1.0
Flask-WTF>=1.2.2
lark>=1.2.2
pandas>=2.3.0